+++ src/index.ts
@@ -77,10 +77,7 @@ export { default as splitThousand } from './splitThousand' // 千分位分割方
 export { default as all } from './all' // 如果所提供的谓词函数对一个集合中的所有元素都返回true，则返回true，否则返回false。
 export { default as any } from './any' // 如果所提供的谓词函数对一个集合中的至少一个元素返回true，则返回true，否则返回false。
 export { default as uuid } from './uuid' // 浏览器端生成uuid，采用v4方法
-export { default as arrayToCSV } from './arrayToCSV' // 将一个二维数组转换为一个逗号分隔的值（CSV）字符串。
-export { default as CSVToArray } from './CSVToArray' // 将一个逗号分隔的值（CSV）字符串转换为一个2D数组。
-export { default as CSVToJSON } from './CSVToJSON' // 将一个逗号分隔的值(CSV)字符串转换为一个2D对象数组。字符串的第一行作为标题行。
-export { default as JSONToCSV } from './JSONToCSV' // 将一个对象数组转换为只包含指定列的逗号分隔值（CSV）字符串。
+export * from './csv' // csv与json、array相互转换
 export { default as RGBToHex } from './RGBToHex' // 将RGB组件的值转换为颜色代码。
 export { default as intersect } from './intersect' // 多个数组求交集
 export { default as union } from './union' // 多个数组求交集